FIX run*.py current dir in PATH
When running a run*.py script (webapp, task, worker) the current directory of
the script was put into sys.path, which was handling folders like
'asl/interface/gearman' as 'gearman' for run_task/tests.

This replaces the current dir in sys.path with ASL dir when run as standalone
scripts. Otherwise it is expecting that ASL is in path