Futurize stage2
I ran the `futurize --stage2` on the project, I did change dictionary
iterators and testing for strings, added map and zip from future
package.

I edited the iterators, using mostly the `iteritems` method from
`future.utils`, because it was mostly in for loops and we used them
wrongly in the firstplace (not using iter/view eg. dict.items()).