Documenation - first draft
I generated a new documentation using sphinx-quickstart. I moved the
api references from index.rst to api.rst and added a getting started
guide.

I also created a README for github, which is a summary of the main page
and gettings started guide. This is also the reason getting_started
is written in markdown (for now).