Full implementation for celery worker
After fixing the gearman worker I changed the TaskQueueWorker interface.
This implements the new Worker interface for the Celery worker with a
proper task execution.

Currently Celery supports only functions as tasks and not methods, so I
implemented one with a worker as injected parameter, but because I
didn't want to put it to global initializers so I created an Injector
Module which is added in run_celery_worker method. This will need a better
solution in the future.

And I improved the Job interface with properties.