FIX click unicode_literals warning
The click library has a warning when using in a module with `from
__future__ import unicode_literals`.

See http://click.pocoo.org/dev/python3/#unicode-literals

They suggest to use the u'' prefix in all the modules using the click,
but this would violate the code style in some files. Using it on all the
code would be too cumbersome. For now I think we can live with the risk,
cli is not the main way to configure the app.