Lazy application initilization
I removed the application initialization code run at importing
`app.application.service_application` and the instance
`service_application`. Now the initialization happens in constructing
the Zsl instance (ServiceApplication).

Removing the global instance meant that

 - all modules which need the app instance are now using DI to get it
 - I changed web performers to factory functions, where the app instance
 is injected
 - in various modules I replaced `app.logger` with `logging`, using the
 logging module directly
 - in run script I construct the app object

I split the initializers into DI modules, which configure the
dependencies, and initializers which are bound into
`InitializationContext`.

To help to maintain a configuration with DI modules I created
`IoCContainer` class which is a named collection of DI modules.