FIX combining policy in resource guard
When guards are applied on top each other, they must combine the
policies from the ancestors. This worked although in reversed order,
because usually you want to add more specific policies onto more
specific resource.

The bug was in guard method wrapping, this is only done once, the first
time the guard is applied and it was pointing only to its original
policies.