Improve the DbTestCase so that it allows for real database unit tests
DbTestCase is suitable for testing data within databases in a way that
first the database is created (schemas of tables). Then the data may be
changed and created and the logic may be tested. Finally all this
happens in a transaction on which rollback is called.

Also in memory databases with sqlite may be used.